# cybercrime/openapi.yaml
components:
  schemas:
...
    ComplaintCreate:
      properties:
        evidence_ids:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Evidence Ids
        incident_description:
          title: Incident Description
          type: string
        internal_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Internal Notes
        investigator_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Investigator Id
        investigator_location_city:
          anyOf:
          - type: string
          - type: 'null'
          title: Investigator Location City
        investigator_location_country:
          anyOf:
          - type: string
          - type: 'null'
          title: Investigator Location Country
        investigator_location_ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Investigator Location Ip
        investigator_location_latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Investigator Location Latitude
        investigator_location_longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Investigator Location Longitude
        officer_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Officer Address
        officer_designation:
          title: Officer Designation
          type: string
        officer_email:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Officer Email
        officer_mobile:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Officer Mobile
        officer_telephone:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Officer Telephone
        wallet_address:
          title: Wallet Address
          type: string
      required:
      - wallet_address
      - officer_designation
      - incident_description
      title: ComplaintCreate
      type: object
    ComplaintResponse:
      properties:
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        evidence_ids:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Evidence Ids
        id:
          title: Id
          type: integer
        incident_description:
          title: Incident Description
          type: string
        internal_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Internal Notes
        investigator_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Investigator Id
        investigator_location_city:
          anyOf:
          - type: string
          - type: 'null'
          title: Investigator Location City
        investigator_location_country:
          anyOf:
          - type: string
          - type: 'null'
          title: Investigator Location Country
        investigator_location_ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Investigator Location Ip
        investigator_location_latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Investigator Location Latitude
        investigator_location_longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Investigator Location Longitude
        officer_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Officer Address
        officer_designation:
          title: Officer Designation
          type: string
        officer_email:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Officer Email
        officer_mobile:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Officer Mobile
        officer_telephone:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Officer Telephone
        status:
          title: Status
          type: string
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
        wallet_address:
          title: Wallet Address
          type: string
      required:
      - id
      - wallet_address
      - investigator_id
      - officer_designation
      - officer_address
      - officer_email
      - officer_mobile
      - officer_telephone
      - incident_description
      - internal_notes
      - evidence_ids
      - investigator_location_city
      - investigator_location_country
      - investigator_location_latitude
      - investigator_location_longitude
      - investigator_location_ip
      - status
      - created_at
      - updated_at
      title: ComplaintResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  description: Backend API for cybercrime investigation dashboard
  title: Cybercrime Investigation API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/complaints/:
    get:
      description: Get all filed complaints
      operationId: get_complaints_api_v1_complaints__get
      parameters:
      - in: query
        name: investigator_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Investigator Id
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ComplaintResponse'
                title: Response Get Complaints Api V1 Complaints  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Complaints
      tags:
      - complaints
    post:
      description: Create a new complaint
      operationId: create_complaint_api_v1_complaints__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Complaint
      tags:
      - complaints
  /api/v1/complaints/{complaint_id}:
    get:
      description: Get a specific complaint by ID
      operationId: get_complaint_api_v1_complaints__complaint_id__get
      parameters:
      - in: path
        name: complaint_id
        required: true
        schema:
          title: Complaint Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Complaint
      tags:
      - complaints
  /api/v1/complaints/{complaint_id}/status:
    patch:
      description: Update complaint status
      operationId: update_complaint_status_api_v1_complaints__complaint_id__status_patch
      parameters:
      - in: path
        name: complaint_id
        required: true
        schema:
          title: Complaint Id
          type: integer
      - in: query
        name: status
        required: true
        schema:
          title: Status
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Complaint Status
      tags:
      - complaints
...
  /api/v1/fraud-predictions/predict:
    post:
      description: Predict fraud for transaction data (without saving to DB)
      operationId: predict_fraud_for_transaction_data_api_v1_fraud_predictions_predict_post
      requestBody:
        content:
          application/json:
            schema:
              title: Transaction
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Predict Fraud For Transaction Data
      tags:
      - fraud-predictions
  /api/v1/fraud-predictions/predict/{transaction_id}:
    get:
      description: Predict fraud for a specific transaction
      operationId: predict_fraud_for_transaction_api_v1_fraud_predictions_predict__transaction_id__get
      parameters:
      - in: path
        name: transaction_id
        required: true
        schema:
          title: Transaction Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Predict Fraud For Transaction
      tags:
      - fraud-predictions
...